refactor(HideLongSummaries): split i18n, constant and core; add index; update LICENSE
