+++ src/HideLongSummaries/modules/constant.ts
@@ -1,0 +1,1 @@
+export const LIMIT = 250;
